+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.16.0
  * Fix 'k' formatting (1 to 24 hours) which incorrectly showed 0 to 23.
  * Tighten up types in a couple of places.
+ * Add dart2js pragmas for inlining to help remove descriptions and other
+   compile-time information from the output.
 
 ## 0.15.8
  * Add return type to some internal methods to improve dart2js output.